Python/YandexMusic-Export/requirements.txt
python-dotenv
yandex-music